syerif20220104-2

+++ clearance_in.js
@@ -1,7 +1,7 @@
 const pool = require('./dbCon');
 const fs = require('fs');
 const path = require('path')
-
+const base_url = process.env.base_url;
 
 const create = (request, response) => {
     const { voyage_id,degree1,minute1,second1,direction1,degree2,minute2,second2,direction2,weather_valid_from,weather_valid_to,weather_data_feed,wind_speed_min,wind_speed_max,wind_from,wind_to,humidity_min,humidity_max,air_pressure,temperature_min,temperature_max,low_tide,high_tide,low_tide_time,high_tide_time,weather,informasi_cuaca_lainnya,speed_kapal,draught_saat_ini,informasi_lainnya,heading_kapal,approval_ksu,dokumen_spm } 
@@ -110,9 +110,42 @@ const read_by_id = (request, response) => {
 
 }
 
+const read_by_voyage_id = (request, response) => {
+
+  const id = parseInt(request.params.id);
+  //console.log('Here');
+  //console.log(id);
+  const {page,rows} = request.body
+  var page_req = page || 1
+  var rows_req = rows || 10
+  var offset = (page_req - 1) * rows_req
+  var res = []
+  var items = []
+
+  pool.query('SELECT count(*) as total FROM tbl_insaf_clearance_in where voyage_id=$1 and is_delete=false', [id], (error, results) => {
+    if (error) {
+      throw error
+    }
+   //console.log(results.rows[0].total)
+   res.push({total:results.rows[0].total})
+
+   var sql= 'SELECT * FROM tbl_insaf_clearance_in where voyage_id=$1 and is_delete=false'
+   pool.query(sql,[id] ,(error, results) => {
+     if (error) {
+       throw error
+     }
+     items.push({rows:results.rows})
+     res.push(items)
+     response.status(200).send({success:true,data:res})
+   })
+
+  })
+
+}
+
 const update = (request, response) => {
     const id = parseInt(request.params.id);
-    const { voyage_id,degree1,minute1,second1,direction1,degree2,minute2,second2,direction2,weather_valid_from,weather_valid_to,weather_data_feed,wind_speed_min,wind_speed_max,wind_from,wind_to,humidity_min,humidity_max,air_pressure,temperature_min,temperature_max,low_tide,high_tide,low_tide_time,high_tide_time,weather,informasi_cuaca_lainnya,heading_kapal,speed_kapal,draught_saat_ini,informasi_lainnya,approval_ksu,dokumen_spm } 
+    const { voyage_id,degree1,minute1,second1,direction1,degree2,minute2,second2,direction2,weather_valid_from,weather_valid_to,weather_data_feed,wind_speed_min,wind_speed_max,wind_from,wind_to,humidity_min,humidity_max,air_pressure,temperature_min,temperature_max,low_tide,high_tide,low_tide_time,high_tide_time,weather,informasi_cuaca_lainnya,heading_kapal,speed_kapal,draught_saat_ini,informasi_lainnya } 
     = request.body;
     let doc;
     //console.log(mmsi);
@@ -134,30 +167,33 @@ const update = (request, response) => {
             throw error
           }
 
-         doc = results.rows[0].dokumen_spm;
-         var doc_path = __dirname +path.join('/dokumens/clearance_in/'+ doc);
-         console.log(doc_path);
-         fs.unlinkSync(doc_path);
-         console.log(doc_path);
+        //  doc = results.rows[0].dokumen_spm;
+        //  var doc_path = __dirname +path.join('/dokumens/clearance_in/'+ doc);
+        //  console.log(doc_path);
+        //  if (fs.existsSync(doc_path)){
+        //   fs.unlinkSync(doc_path);
+        //  }
+         
+        //  console.log(doc_path);
 
-         let sampleFile = request.files.dokumen_spm;
-         console.log(sampleFile);
-          const now = Date.now()
-          let name = now + '_' + sampleFile['name'].replace(/\s+/g, '')
-          console.log(__dirname);
-          sampleFile.mv(path.join(__dirname + '/dokumens/clearance_in/') + name, function (err) {
-              if (err){
-                console.log(err);
-              }
+        //  let sampleFile = request.files.dokumen_spm;
+        //  console.log(sampleFile);
+        //   const now = Date.now()
+        //   let name = now + '_' + sampleFile['name'].replace(/\s+/g, '')
+        //   console.log(__dirname);
+        //   sampleFile.mv(path.join(__dirname + '/dokumens/clearance_in/') + name, function (err) {
+        //       if (err){
+        //         console.log(err);
+        //       }
                   
-          });
+        //   });
 
-          console.log(name);
+        //   console.log(name);
          const update_time = new Date;
-         pool.query('UPDATE tbl_insaf_clearance_in SET voyage_id=$1,degree1=$2,minute1=$3,second1=$4,direction1=$5,degree2=$6,minute2=$7,second2=$8,direction2=$9,weather_valid_from=$10,weather_valid_to=$11,weather_data_feed=$12,wind_speed_min=$13,wind_speed_max=$14,wind_from=$15,wind_to=$16,humidity_min=$17,humidity_max=$18,air_pressure=$19,temperature_min=$20,temperature_max=$21,low_tide=$22,high_tide=$23,low_tide_time=$24,high_tide_time=$25,weather=$26,informasi_cuaca_lainnya=$27,heading_kapal=$28,speed_kapal=$29,draught_saat_ini=$30,informasi_lainnya=$31,approval_ksu=$32,dokumen_spm=$33,updated_at=$34 where id=$35'
-         , [voyage_id,degree1,minute1,second1,direction1,degree2,minute2,second2,direction2,weather_valid_from,weather_valid_to,weather_data_feed,wind_speed_min,wind_speed_max,wind_from,wind_to,humidity_min,humidity_max,air_pressure,temperature_min,temperature_max,low_tide,high_tide,low_tide_time,high_tide_time,weather,informasi_cuaca_lainnya,heading_kapal,speed_kapal,draught_saat_ini,informasi_lainnya,approval_ksu,name,update_time,id], (error, results) =>{
+         pool.query('UPDATE tbl_insaf_clearance_in SET voyage_id=$1,degree1=$2,minute1=$3,second1=$4,direction1=$5,degree2=$6,minute2=$7,second2=$8,direction2=$9,weather_valid_from=$10,weather_valid_to=$11,weather_data_feed=$12,wind_speed_min=$13,wind_speed_max=$14,wind_from=$15,wind_to=$16,humidity_min=$17,humidity_max=$18,air_pressure=$19,temperature_min=$20,temperature_max=$21,low_tide=$22,high_tide=$23,low_tide_time=$24,high_tide_time=$25,weather=$26,informasi_cuaca_lainnya=$27,heading_kapal=$28,speed_kapal=$29,draught_saat_ini=$30,informasi_lainnya=$31,updated_at=$32 where id=$33'
+         , [voyage_id,degree1,minute1,second1,direction1,degree2,minute2,second2,direction2,weather_valid_from,weather_valid_to,weather_data_feed,wind_speed_min,wind_speed_max,wind_from,wind_to,humidity_min,humidity_max,air_pressure,temperature_min,temperature_max,low_tide,high_tide,low_tide_time,high_tide_time,weather,informasi_cuaca_lainnya,heading_kapal,speed_kapal,draught_saat_ini,informasi_lainnya,update_time,id], (error, results) =>{
            if (error) {
-             // throw error
+              throw error
              //response.status(201).send(error)
              //console.log(error);
              if (error.code == '23505')
@@ -208,13 +244,17 @@ const update_ksu = (request, response) => {
          doc = results.rows[0].dokumen_spm;
          var doc_path = __dirname +path.join('/dokumens/clearance_in/'+ doc);
          console.log(doc_path);
-         fs.unlinkSync(doc_path);
+         if (fs.existsSync(doc_path)){
+          fs.unlinkSync(doc_path);
+         }
+         
          console.log(doc_path);
 
          let sampleFile = request.files.dokumen_spm;
          console.log(sampleFile);
           const now = Date.now()
           let name = now + '_' + sampleFile['name'].replace(/\s+/g, '')
+          var complete_path = base_url+'dokumens/clearance_in/'+name;
           console.log(__dirname);
           sampleFile.mv(path.join(__dirname + '/dokumens/clearance_in/') + name, function (err) {
               if (err){
@@ -225,10 +265,10 @@ const update_ksu = (request, response) => {
 
           console.log(name);
          const update_time = new Date;
-         pool.query('UPDATE tbl_insaf_clearance_in SET voyage_id=$1,degree1=$2,minute1=$3,second1=$4,direction1=$5,degree2=$6,minute2=$7,second2=$8,direction2=$9,weather_valid_from=$10,weather_valid_to=$11,weather_data_feed=$12,wind_speed_min=$13,wind_speed_max=$14,wind_from=$15,wind_to=$16,humidity_min=$17,humidity_max=$18,air_pressure=$19,temperature_min=$20,temperature_max=$21,low_tide=$22,high_tide=$23,low_tide_time=$24,high_tide_time=$25,weather=$26,informasi_cuaca_lainnya=$27,heading_kapal=$28,speed_kapal=$29,draught_saat_ini=$30,informasi_lainnya=$31,approval_ksu=$32,dokumen_spm=$33,updated_at=$34 where id=$35'
-         , [voyage_id,degree1,minute1,second1,direction1,degree2,minute2,second2,direction2,weather_valid_from,weather_valid_to,weather_data_feed,wind_speed_min,wind_speed_max,wind_from,wind_to,humidity_min,humidity_max,air_pressure,temperature_min,temperature_max,low_tide,high_tide,low_tide_time,high_tide_time,weather,informasi_cuaca_lainnya,heading_kapal,speed_kapal,draught_saat_ini,informasi_lainnya,approval_ksu,name,update_time,id], (error, results) =>{
+         pool.query('UPDATE tbl_insaf_clearance_in SET approval_ksu=$1,dokumen_spm=$2,updated_at=$3,url_dokumen_spm=$4 where id=$5'
+         , [approval_ksu,name,update_time,complete_path,id], (error, results) =>{
            if (error) {
-             // throw error
+              throw error
              //response.status(201).send(error)
              //console.log(error);
              if (error.code == '23505')
@@ -302,11 +342,25 @@ const delete_ = (request, response) => {
     
 }
 
+
+
+const download = (request, response) => {
+  const filename = request.params.filename;
+  console.log(filename);
+  var doc_path = __dirname + path.join('/dokumens/clearance_in/'+ filename);
+  console.log(doc_path);
+  response.download(doc_path);
+  //response.status(200).send({success:true,data:'data berhasil diunduh'})
+};
+
+
 module.exports = {
     create,
     read,
     read_by_id,
+    read_by_voyage_id,
     update,
     update_ksu,
     delete_,
+    download,
     }
